Create  Websites
This program prompts the user for a business name, removes all spaces and converts the name to title case, and then suggests a web address by adding www. to the front of the name and .com to the end. Note that the program assumes that the web address should be in title case (i.e., with the first letter of each word capitalized) and lower case (i.e., with all other letters in lower case). This is a common convention for web addresses.